improve optical pumping
pumping rate is now proportional to population difference between the ground state and the excitation level
documentation updated
ATF models updated
