--- hitran_data/constants_626_gbd_ef_tied.csv
@@ -1,59 +1,0 @@
-ef,m1,m2,l2,m3,r,G_Hz,B_Hz,D_Hz
-e,0,0,0,0,1,0.000000000000e+00,1.169631002439e+10,4.046504307019e+03
-e,0,0,0,1,1,7.042657113846e+13,1.160500798336e+10,4.118117804216e+03
-f,0,0,0,1,1,7.042657113846e+13,1.161046100106e+10,4.071751075489e+03
-e,0,0,0,2,1,1.401036221818e+14,1.151359940461e+10,4.124218376200e+03
-f,0,0,0,2,1,1.401036221818e+14,1.151802770455e+10,4.226629359945e+03
-e,0,0,0,3,1,2.090338876726e+14,1.142058406146e+10,4.132377085977e+03
-e,0,1,1,0,1,2.000870723423e+13,1.170883229596e+10,3.787898625981e+03
-f,0,1,1,0,1,2.000870723423e+13,1.173355147250e+10,4.344731303061e+03
-e,0,1,1,1,1,9.005861532156e+13,1.162341199352e+10,4.134888446701e+03
-f,0,1,1,1,1,9.005861532156e+13,1.163601416108e+10,3.977154744082e+03
-e,0,2,2,0,1,4.002697976998e+13,1.173976865203e+10,4.180103323748e+03
-f,0,2,2,0,1,4.002697976998e+13,1.174639358604e+10,4.266919199761e+03
-e,0,2,2,1,1,1.097027227778e+14,1.165128915689e+10,4.394535201164e+03
-f,0,2,2,1,1,1.097027227778e+14,1.165259124379e+10,3.996689816421e+03
-e,0,3,3,0,1,6.005657718135e+13,1.176280453443e+10,4.319480207673e+03
-f,0,3,3,0,1,6.005657718135e+13,1.176511163147e+10,4.187980989287e+03
-e,0,3,3,1,1,1.293584377454e+14,1.167358328571e+10,4.316196360383e+03
-f,0,3,3,1,1,1.293584377454e+14,1.167546944615e+10,4.201953154688e+03
-e,0,4,4,0,1,8.009642186046e+13,1.178549509189e+10,4.568755385220e+03
-f,0,4,4,0,1,8.009642186046e+13,1.178663257331e+10,4.290462561423e+03
-e,0,4,4,1,1,1.490251516469e+14,1.169485245441e+10,4.338103023116e+03
-f,0,4,4,1,1,1.490251516469e+14,1.169854170212e+10,4.301163475657e+03
-e,1,0,0,0,1,4.161758350803e+13,1.169586961805e+10,3.507599994277e+03
-f,1,0,0,0,1,4.161758350803e+13,1.170138762838e+10,3.630922256225e+03
-e,1,0,0,0,2,3.853711871365e+13,1.170368745922e+10,4.616584175571e+03
-f,1,0,0,0,2,3.853711871365e+13,1.170990734862e+10,4.866552018082e+03
-e,1,0,0,1,1,1.113671166998e+14,1.160344530680e+10,3.574265097349e+03
-f,1,0,0,1,1,1.113671166998e+14,1.160642237227e+10,3.743659850806e+03
-e,1,0,0,1,2,1.083110820941e+14,1.161655583287e+10,4.835351425765e+03
-f,1,0,0,1,2,1.083110820941e+14,1.161924787683e+10,4.945501471085e+03
-e,1,1,1,0,1,6.226236958537e+13,1.170563488149e+10,3.346084825837e+03
-f,1,1,1,0,1,6.226236958537e+13,1.173674926071e+10,4.508383118119e+03
-e,1,1,1,0,2,5.793445681594e+13,1.171567093453e+10,4.122914495480e+03
-f,1,1,1,0,2,5.793445681594e+13,1.174563427665e+10,5.248555214101e+03
-e,1,1,1,1,1,1.316277760735e+14,1.162259150690e+10,4.286383716313e+03
-f,1,1,1,1,1,1.316277760735e+14,1.163187407488e+10,3.032147378428e+03
-e,1,1,1,1,2,1.273436061494e+14,1.163525010997e+10,5.077450685075e+03
-f,1,1,1,1,2,1.273436061494e+14,1.164405718942e+10,3.901068509167e+03
-e,1,2,2,0,1,8.276502082560e+13,1.173828471657e+10,4.112208505820e+03
-f,1,2,2,0,1,8.276502082560e+13,1.174065388296e+10,4.151577853655e+03
-e,1,2,2,0,2,7.749740868503e+13,1.175049137032e+10,4.324518672173e+03
-f,1,2,2,0,2,7.749740868503e+13,1.175295661241e+10,4.774179983318e+03
-e,1,2,2,1,1,1.517490112067e+14,1.164892937988e+10,4.558204137025e+03
-f,1,2,2,1,1,1.517490112067e+14,1.164720435144e+10,3.456892000050e+03
-e,1,2,2,1,2,1.465385278435e+14,1.166132029768e+10,4.552897698885e+03
-f,1,2,2,1,2,1.465385278435e+14,1.166162847759e+10,4.302987013066e+03
-e,2,0,0,0,1,8.385697847610e+13,1.170912146687e+10,2.604466457395e+03
-f,2,0,0,0,1,8.385697847610e+13,1.171364235010e+10,3.333550201153e+03
-e,2,0,0,0,2,8.008003638671e+13,1.167742332601e+10,3.653555164528e+03
-f,2,0,0,0,2,8.008003638671e+13,1.168231248072e+10,4.583184676389e+03
-e,2,0,0,0,3,7.639902976833e+13,1.172419591329e+10,5.105783367917e+03
-f,2,0,0,0,3,7.639902976833e+13,1.172926632876e+10,5.981494679295e+03
-e,2,0,0,1,1,1.528853378836e+14,1.161665813688e+10,3.017715444721e+03
-f,2,0,0,1,1,1.528853378836e+14,1.161239381092e+10,1.348715648267e+03
-e,2,0,0,1,2,1.492316393729e+14,1.158982265559e+10,4.440225692646e+03
-f,2,0,0,1,2,1.492316393729e+14,1.158712714902e+10,5.011623140410e+03
-e,2,0,0,1,3,1.455095934043e+14,1.163696599692e+10,5.424181137832e+03
-f,2,0,0,1,3,1.455095934043e+14,1.163470333175e+10,4.393024368093e+03
