--- hitran_data/constants_626_gbd_nonlinear.csv
@@ -1,31 +1,0 @@
-ef,m1,m2,l2,m3,r,G_Hz,B_Hz,D_Hz
-e,0,0,0,0,1,0.000000000000e+00,1.169627094143e+10,4.041954788872e+03
-e,0,0,0,1,1,7.042622256897e+13,1.160509870406e+10,4.121864717521e+03
-e,0,0,0,2,1,1.401037984494e+14,1.151349659071e+10,4.114930091268e+03
-e,0,0,0,3,1,2.090339780221e+14,1.142052064431e+10,4.126352013639e+03
-e,0,1,1,0,1,2.000988007928e+13,1.170840845382e+10,3.758890923225e+03
-e,0,1,1,1,1,9.005816740044e+13,1.162356214362e+10,4.143720899089e+03
-e,0,2,2,0,1,4.002809246543e+13,1.173928652789e+10,4.142138961226e+03
-e,0,2,2,1,1,1.097023004059e+14,1.165143102068e+10,4.403657176689e+03
-e,0,3,3,0,1,6.005566125228e+13,1.176320245835e+10,4.355697218110e+03
-e,0,3,3,1,1,1.293585906308e+14,1.167354266356e+10,4.317600989964e+03
-e,0,4,4,0,1,8.009604715989e+13,1.178572946598e+10,4.596384317681e+03
-e,0,4,4,1,1,1.490256358272e+14,1.169463492630e+10,4.322209018477e+03
-e,1,0,0,0,1,4.161756743950e+13,1.169583725427e+10,3.503549186969e+03
-e,1,0,0,0,2,3.853711469951e+13,1.170365211315e+10,4.612581377225e+03
-e,1,0,0,1,1,1.113668289271e+14,1.160351682092e+10,3.577525353063e+03
-e,1,0,0,1,2,1.083108027543e+14,1.161663207469e+10,4.839129538030e+03
-e,1,1,1,0,1,6.226296061896e+13,1.170535091534e+10,3.321596375166e+03
-e,1,1,1,0,2,5.793455517267e+13,1.171561063644e+10,4.117351114780e+03
-e,1,1,1,1,1,1.316257855303e+14,1.162353703602e+10,4.366059393021e+03
-e,1,1,1,1,2,1.273412648554e+14,1.163640448284e+10,5.175995848629e+03
-e,1,2,2,0,1,8.276505499955e+13,1.173825240317e+10,4.110650182532e+03
-e,1,2,2,0,2,7.749787690626e+13,1.175022136804e+10,4.298639399791e+03
-e,1,2,2,1,1,1.517491478540e+14,1.164879239871e+10,4.544363877122e+03
-e,1,2,2,1,2,1.465390713382e+14,1.166097185398e+10,4.518779360566e+03
-e,2,0,0,0,1,8.385710860621e+13,1.170902224059e+10,2.596012724810e+03
-e,2,0,0,0,2,8.008020709394e+13,1.167729419795e+10,3.640350067575e+03
-e,2,0,0,0,3,7.639926692139e+13,1.172403824702e+10,5.091119947169e+03
-e,2,0,0,1,1,1.528851952887e+14,1.161667112602e+10,3.017986649025e+03
-e,2,0,0,1,2,1.492314172105e+14,1.158985893964e+10,4.441594091475e+03
-e,2,0,0,1,3,1.455095508960e+14,1.163693482627e+10,5.420480508995e+03
