--- hitran_data/constants_737.csv
@@ -1,21 +1,0 @@
-ef,m1,m2,l2,m3,r,G_Hz,B_Hz,D_Hz
-e,0,0,0,0,1,0.000000000000e+00,1.100702644344e+10,3.576460644048e+03
-e,0,0,0,1,1,6.788379115733e+13,1.092353906579e+10,3.555627286159e+03
-e,0,1,1,0,1,1.927681072980e+13,1.102100304442e+10,2.198413436444e+03
-f,0,1,1,0,1,1.927681072980e+13,1.103692982585e+10,3.651384017743e+03
-e,0,1,1,1,1,8.681313922369e+13,1.093866701631e+10,2.330733888069e+03
-f,0,1,1,1,1,8.681313922369e+13,1.095365118387e+10,3.552910236519e+03
-e,0,2,2,0,1,3.856357643368e+13,1.104905926368e+10,2.367140957152e+03
-f,0,2,2,0,1,3.856357643368e+13,1.104788575007e+10,3.663089387984e+03
-e,0,2,2,1,1,1.057530745582e+14,1.096702433598e+10,2.510324904534e+03
-f,0,2,2,1,1,1.057530745582e+14,1.096570801040e+10,3.666900967906e+03
-e,1,0,0,0,1,4.020966834273e+13,1.100627733219e+10,4.907868707063e+03
-f,1,0,0,0,1,4.020966834273e+13,1.100485847027e+10,3.374641947544e+03
-e,1,0,0,0,2,3.729075642952e+13,1.102151848489e+10,6.265747416225e+03
-f,1,0,0,0,2,3.729075642952e+13,1.101973842627e+10,4.608183644789e+03
-e,1,0,0,1,1,1.074725498325e+14,1.091919840943e+10,3.698303862338e+03
-e,1,0,0,1,2,1.045616956346e+14,1.093811516709e+10,4.958588504908e+03
-e,1,1,1,1,1,1.269873291161e+14,1.093008950180e+10,2.283653599774e+03
-f,1,1,1,1,1,1.269873291161e+14,1.095237105615e+10,3.308898225946e+03
-e,1,1,1,1,2,1.229318008679e+14,1.094573558383e+10,3.639118785068e+03
-f,1,1,1,1,2,1.229318008679e+14,1.097045254678e+10,4.442354812157e+03
